fix: normalize avgjudge sentinel to i32::MAX on DB write for Java compat
The in-memory sentinel i64::MAX was written directly to the SQLite DB,
breaking Java clients that read the column as a 32-bit int. Now
normalizes i64::MAX back to i32::MAX on write, matching the original
Java storage format.

+++ crates/rubato-core/src/score_database_accessor/tests.rs
@@ -238,9 +238,10 @@ fn test_score_data_to_value_default_fields() {
         score_data_to_value(&sd, "minbp"),
         rusqlite::types::Value::Integer(i32::MAX as i64)
     );
+    // score_data_to_value normalizes i64::MAX sentinel to i32::MAX for Java DB compatibility
     assert_eq!(
         score_data_to_value(&sd, "avgjudge"),
-        rusqlite::types::Value::Integer(i64::MAX)
+        rusqlite::types::Value::Integer(i32::MAX as i64)
     );
     assert_eq!(
         score_data_to_value(&sd, "seed"),
